register: display error and error message in jaeger
Currently when a functions fails either due to an error or panic, jaeger
has no way to display such information. It creates confusion in the
traces that we don't know if a function has succeded or not.
This PR adds a `error` tag to jaeger trace which jaeger UI is able to
parse and display.

Change-Id: I7165df475d1d5c4a619ca6dc8401ed3765622323